backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
django-cors-headers==4.4.0
